+++ packages/eslint-config/rules/consistent-default-export-name.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ['plugin:consistent-default-export-name/fixed'],
+}