+++ CHANGELOG.md
@@ -4,11 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+- @xrplf/eslint-config lints filenames such that they match their default export
+- @xrplf/eslint-config's engines and peer dependencies are looser
+- @xrplf/eslint-config allows the `category` doc tag for use with type doc
+
 ## 1.3.0
 - @xrplf/eslint-config allows peers with typescript 3 or 4
 - @xrplf/eslint-config skips blank lines and comments in
   `max-lines` and `max-lines-per-function` rules
-
 
 ## 1.2.2
 First version that is correctly tagged with this CHANGELOG. Previous releases